feat: in-game에 필요한 dto 생성

+++ ingame-search/src/main/java/sejong/teemo/ingamesearch/dto/Champion.java
@@ -1,9 +1,9 @@
 package sejong.teemo.ingamesearch.dto;
 
 public record Champion(
-        int kills,
-        int deaths,
-        int assists,
+        double kills,
+        double deaths,
+        double assists,
         double rating,
         int recentGameCount
 ) {
